the-KINGV2-main/backend/utils/__init__.py
"""LittleHelper AI Utilities"""
import hashlib
import bcrypt
import jwt
from datetime import datetime, timezone, timedelta
from fastapi import HTTPException, Header
from typing import Optional
import os

# JWT Configuration
JWT_SECRET = os.environ.get('JWT_SECRET', 'littlehelper-ai-secret-key-2024')
JWT_ALGORITHM = "HS256"
JWT_EXPIRATION_HOURS = 24


def hash_password(password: str) -> str:
    """Hash a password using bcrypt"""
    return bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt()).decode('utf-8')


def verify_password(password: str, hashed: str) -> bool:
    """Verify a password against its hash"""
    return bcrypt.checkpw(password.encode('utf-8'), hashed.encode('utf-8'))


def create_token(user_id: str, role: str) -> str:
    """Create a JWT token"""
    payload = {
        "user_id": user_id,
        "role": role,
        "exp": datetime.now(timezone.utc) + timedelta(hours=JWT_EXPIRATION_HOURS)
    }
    return jwt.encode(payload, JWT_SECRET, algorithm=JWT_ALGORITHM)


def decode_token(token: str) -> dict:
    """Decode and validate a JWT token"""
    try:
        return jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token expired")
    except jwt.InvalidTokenError:
        raise HTTPException(status_code=401, detail="Invalid token")


def normalize_question(question: str) -> str:
    """Normalize a question for caching"""
    return ' '.join(question.lower().strip().split())


def hash_question(question: str) -> str:
    """Create a hash of a normalized question"""
    return hashlib.sha256(normalize_question(question).encode()).hexdigest()


def estimate_tokens(text: str) -> int:
    """Rough estimate of tokens for text"""
    return int(len(text.split()) * 1.3)
